Splitting static functions from Array to individual files
New Classes / files created include:
- Data
- Arith
- Signal
- Algorithm
